refactor: export single handler + utils
