+++ src/rollup/config.ts
@@ -205,7 +205,7 @@ export const getRollupConfig = (nitro: Nitro) => {
     entries: resolveAliases({
       '#nitro': runtimeDir,
       '#config': resolve(runtimeDir, 'config'),
-      '#nitro-error': resolve(runtimeDir, 'error/handler'),
+      '#nitro-error': resolve(runtimeDir, 'error'),
       '#_config': resolve(runtimeDir, 'config'),
       '#paths': resolve(runtimeDir, 'paths'),
       '#cache': resolve(runtimeDir, 'cache'),
