--- src/types/error.ts
@@ -1,12 +1,0 @@
-export interface ErrorPayload {
-  error: any
-  isJsonRequest: boolean
-  stack: Array<{ text: string, internal: boolean }>
-  errorObject: {
-    url?: string
-    statusCode: number
-    statusMessage: string
-    message: string
-    description: string
-  }
-}
