+++ src/types/index.ts
@@ -1,5 +1,4 @@
 // import './shims' TODO
-export * from './error'
 export * from './fetch'
 export * from './nitro'
 export * from './handler'
